+++ src/OpenTelemetry/CHANGELOG.md
@@ -22,6 +22,10 @@ Notes](../../RELEASENOTES.md).
 * Update `OpenTelemetrySdkEventSource` to support the W3C randomness flag.
   ([#7301](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7301))
 
+* **Breaking Change** Explicit histogram boundaries no longer allow more than
+  10 million values.
+  ([#7165](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7165))
+
 ## 1.15.3
 
 Released 2026-Apr-21